Backend/requirement.txt
fastapi 
uvicorn 
pydantic
typing-extensions
supabase
requests
python-dotenv
pydantic[email]
python-multipart
pandas